# api-server/routes/admin.ts
import { Router } from "express";
import { db, usersTable, botsTable, announcementsTable } from "@workspace/db";
import { eq } from "drizzle-orm";
import crypto from "crypto";
import { requireAdmin } from "./auth";

const router = Router();

// GET /api/admin/users
router.get("/admin/users", requireAdmin, async (req: any, res) => {
  try {
    const users = await db.select().from(usersTable);
    const usersWithBotCount = await Promise.all(users.map(async (u) => {
      const bots = await db.select().from(botsTable).where(eq(botsTable.userId, u.id));
      return {
        id: u.id,
        name: u.name,
        email: u.email,
        avatar: u.avatar,
        role: u.role,
        plan: u.plan,
        status: u.status,
        botCount: bots.length,
        lastLogin: u.lastLogin?.toISOString() ?? null,
        createdAt: u.createdAt.toISOString(),
      };
    }));
    res.json(usersWithBotCount);
  } catch (err) {
    req.log.error({ err }, "Admin list users error");
    res.status(500).json({ error: "Internal server error" });
  }
});

// PATCH /api/admin/users/:userId
router.patch("/admin/users/:userId", requireAdmin, async (req: any, res) => {
  try {
    const { role, status, plan } = req.body;
    const update: Record<string, any> = {};
    if (role !== undefined) update.role = role;
    if (status !== undefined) update.status = status;
    if (plan !== undefined) update.plan = plan;
    const [user] = await db.update(usersTable).set(update)
      .where(eq(usersTable.id, req.params.userId))
      .returning();
    if (!user) {
      res.status(404).json({ error: "User not found" });
      return;
    }
    const bots = await db.select().from(botsTable).where(eq(botsTable.userId, user.id));
    res.json({
      id: user.id,
      name: user.name,
      email: user.email,
      avatar: user.avatar,
      role: user.role,
      plan: user.plan,
      status: user.status,
      botCount: bots.length,
      lastLogin: user.lastLogin?.toISOString() ?? null,
      createdAt: user.createdAt.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Admin update user error");
    res.status(500).json({ error: "Internal server error" });
  }
});

// DELETE /api/admin/users/:userId
router.delete("/admin/users/:userId", requireAdmin, async (req: any, res) => {
  try {
    await db.delete(usersTable).where(eq(usersTable.id, req.params.userId));
    res.status(204).end();
  } catch (err) {
    req.log.error({ err }, "Admin delete user error");
    res.status(500).json({ error: "Internal server error" });
  }
});

// GET /api/admin/stats
router.get("/admin/stats", requireAdmin, async (req: any, res) => {
  try {
    const users = await db.select().from(usersTable);
    const bots = await db.select().from(botsTable);
    const today = new Date();
    today.setHours(0, 0, 0, 0);
    const newUsersToday = users.filter(u => u.createdAt >= today).length;
    const totalMessages = bots.reduce((acc, b) => acc + b.messageCount, 0);

    const planBreakdown = ["free", "starter", "pro", "enterprise"].map(plan => ({
      plan,
      count: users.filter(u => u.plan === plan).length,
    }));

    const revenueByMonth = Array.from({ length: 6 }, (_, i) => {
      const d = new Date();
      d.setMonth(d.getMonth() - (5 - i));
      return {
        month: d.toLocaleString("default", { month: "short" }),
        revenue: Math.floor(Math.random() * 5000) + 1000,
      };
    });

    res.json({
      totalUsers: users.length,
      totalBots: bots.length,
      totalRevenue: 12500,
      activeUsers: users.filter(u => u.status === "active").length,
      newUsersToday,
      totalMessages,
      revenueByMonth,
      planBreakdown,
    });
  } catch (err) {
    req.log.error({ err }, "Admin stats error");
    res.status(500).json({ error: "Internal server error" });
  }
});

// GET /api/admin/announcements
router.get("/admin/announcements", requireAdmin, async (req: any, res) => {
  try {
    const items = await db.select().from(announcementsTable);
    res.json(items.map(a => ({
      id: a.id,
      title: a.title,
      message: a.message,
      type: a.type,
      createdAt: a.createdAt.toISOString(),
    })));
  } catch (err) {
    req.log.error({ err }, "List announcements error");
    res.status(500).json({ error: "Internal server error" });
  }
});

// POST /api/admin/announcements
router.post("/admin/announcements", requireAdmin, async (req: any, res) => {
  try {
    const { title, message, type } = req.body;
    const [announcement] = await db.insert(announcementsTable).values({
      id: crypto.randomUUID(),
      title,
      message,
      type: type ?? "info",
    }).returning();
    res.status(201).json({
      id: announcement.id,
      title: announcement.title,
      message: announcement.message,
      type: announcement.type,
      createdAt: announcement.createdAt.toISOString(),
    });
  } catch (err) {
    req.log.error({ err }, "Create announcement error");
    res.status(500).json({ error: "Internal server error" });
  }
});

// DELETE /api/admin/announcements/:announcementId
router.delete("/admin/announcements/:announcementId", requireAdmin, async (req: any, res) => {
  try {
    await db.delete(announcementsTable).where(eq(announcementsTable.id, req.params.announcementId));
    res.status(204).end();
  } catch (err) {
    req.log.error({ err }, "Delete announcement error");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;