# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IrForge - Telegram Bot Builder Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Users
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/password:
    patch:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChange"
      responses:
        "200":
          description: OK

  # Dashboard summary
...
  /themes:
    get:
      operationId: listThemes
      tags: [themes]
      summary: List available themes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Theme"
    post:
      operationId: createTheme
      tags: [themes]
      summary: Create a custom theme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
  /themes/{themeId}:
    patch:
      operationId: updateTheme
      tags: [themes]
      summary: Update a theme
      parameters:
        - name: themeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"
    delete:
      operationId: deleteTheme
      tags: [themes]
      summary: Delete a theme
      parameters:
        - name: themeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /themes/apply:
    post:
      operationId: applyTheme
      tags: [themes]
      summary: Apply a theme as the active theme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyThemeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"

  # Admin
...
  /admin/users/{userId}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Admin - update a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin - delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/announcements:
    get:
      operationId: listAnnouncements
      tags: [admin]
      summary: List announcements
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [admin]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /admin/announcements/{announcementId}:
    delete:
      operationId: deleteAnnouncement
      tags: [admin]
      summary: Delete an announcement
      parameters:
        - name: announcementId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, plan, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        botCount:
          type: integer
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: string
        bio:
          type: string
        telegramUsername:
          type: string
    PasswordChange:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    # Dashboard
...
    Theme:
      type: object
      required: [id, name, mode, primaryColor, backgroundColor, userId, isDefault, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        mode:
          type: string
          enum: [light, dark, system]
        primaryColor:
          type: string
        backgroundColor:
          type: string
        foregroundColor:
          type: string
        accentColor:
          type: string
        borderRadius:
          type: string
        fontFamily:
          type: string
        userId:
          type: string
        isDefault:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
    ThemeInput:
      type: object
      required: [name, mode, primaryColor]
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [light, dark, system]
        primaryColor:
          type: string
        backgroundColor:
          type: string
        foregroundColor:
          type: string
        accentColor:
          type: string
        borderRadius:
          type: string
        fontFamily:
          type: string
    ThemeUpdate:
      type: object
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [light, dark, system]
        primaryColor:
          type: string
        backgroundColor:
          type: string
        foregroundColor:
          type: string
        accentColor:
          type: string
        borderRadius:
          type: string
        fontFamily:
          type: string
    ApplyThemeInput:
      type: object
      required: [themeId]
      properties:
        themeId:
          type: string

    # Admin
    AdminUser:
      type: object
      required: [id, name, email, role, plan, status, botCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        avatar:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        plan:
          type: string
        status:
          type: string
          enum: [active, suspended, banned]
        botCount:
          type: integer
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, banned]
        plan:
          type: string
...
    Announcement:
      type: object
      required: [id, title, message, type, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, error]
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, message, type]
      properties:
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, warning, success, error]